services/workload_models/pyproject.toml
[project]
name = "workload-models"
version="0.1.0"
requires-python = ">=3.12"

dependencies = [
    "pyzmq",
    "scikit-learn",
    "pandas",
    "matplotlib",
    "numpy",
    "debugpy"
]

[dependency-groups]
dev = [
    "pytest",
    "black",
    "mypy",
]

[project.scripts]
workload_model = "workload_models.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"